Move migrations from update handler to MigrationService
